Implement RandomHostPolicy
